# Faster TensorCPU KNN Search.