Changes:
- Assuming `num_neighbors = min(dataset_size, knn)`, removed checks. 
- Replaced `Filled tensor of {-1} of shape {num_query_points, knn}` with `Empty tensor of shape {num_query_points, num_neighbors}`.
----
Effect on performance of `estimate_noramls`:

```
------------------------------------------------------------------------------------------------
Benchmark                                                      Before             After   
------------------------------------------------------------------------------------------------
EstimateNormals/CPU F32 KNN[0.02 | 30]                      51.8 ms         23.4 ms          
EstimateNormals/CPU F64 KNN[0.02 | 30]                      54.6 ms         24.1 ms          
LegacyEstimateNormals/Legacy KNN[0.02 | 30]                 27.4 ms         27.4 ms
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/isl-org/open3d/3945)
<!-- Reviewable:end -->